implement arithmetic right shift (#76)
* implement arithmetic right shift

* workaround Nim VM 'cast' limitation

* fix high(stint) bug

* fix compile time bit shift bug

* add test for compile time shift and high(stint)

* add tests against ttmath

+++ tests/ttmath_compat.nim
@@ -9,11 +9,11 @@ template asSt*(val: Int): auto =
   type TargetType = StInt[val.NumBits]
   cast[ptr TargetType](unsafeAddr val)[]
 
-template asTT*[N: static[int]](arr: array[N, uint]): auto =
+template asTT*[N: static[int]](arr: array[N, uint64]): auto =
   type TargetType = UInt[N * 64]
   cast[ptr TargetType](unsafeAddr arr[0])[]
 
-template asTT*[N: static[int]](arr: array[N, int]): auto =
+template asTT*[N: static[int]](arr: array[N, int64]): auto =
   type TargetType = Int[N * 64]
   cast[ptr TargetType](unsafeAddr arr[0])[]
